cli: Move visit_collapsed_untracked_paths to cli_util
We will introduce similar method for tracked paths, which would need to
be in cli_util, so better have this one in there too.
